% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCor.R
\name{computeC}
\alias{computeC}
\title{Automatically compute the  correlation between genes and clinical features}
\usage{
computeC(
  gene_expression,
  clinical_feature,
  "lymph"
)  
}
\description{
Automatically compute the  correlation between genes and clinical features. For further information on requirements as well as how to implement this tool, please visit my Github repository: https://github.com/huynguyen250896/computeC
}
\examples{
computeC(gene_expression,clinical_feature,"lymph")
}
